Trait any_ref::Type

pub unsafe trait Type<'a> {
    type Static: ?Sized + 'static;
}

Trait which provides TypeId for types which are not static.

To be safe this trait requires that Static is the same type as Self but with the lifetimes replaced with 'static

Associated Types

type Static: ?Sized + 'static

Type which should be the same as Self but with the static lifetime