Module any_ref::type_subtitute

Internal traits and their implementation for &T and &mut T. It is not recommended to manually implement traits here and there is no need to know this traits for practical use. Please use make_any_ref macro if possible.

Structs

Reference

ReferenceMut

Traits

LifetimeDowncast

The target type of the struct implemented ReturnType could “downgrade” its lifetime from higher to lower. Generally, this function may not requires more complex operation than return itself directly.

ReturnType

Indicates that a struct is exist to return other type with any lifetime